Load the openpgp.js code asynchronously
Need to script-inject a script because we need to load them in given order:
- load openpgp.js
- enable the "encrypt" button on-load
- the on load handler requires jQuery

And since this is async world now we can't guarantee the order any other way, I think. I could use require.js but that would be another 15k of code. This way we know that jQuery is loaded before the handler is added and possibly executed.