First shot at a blog, just routing and displaying a single post
The routing is bit complicated because the post will be directly in domain root but we already have some presenters in root, too. So we have a special route which checks if the given URL is a blogpost and if not it just returns null which means it's ignored and other routes are explored.