Add (empty) form handler to Pulse search
Because Nette now throws "Form was submitted but there are no associated handlers." when there are no onSuccess handlers attached. But we don't need any because the form only "moves" the values from input fields to the URL and then the action method takes it from there. Params can also be added to the URL manually, not via the form, so the code needs to handle search and filters not in the onSuccess handler because that might not be even called (when passing params manually in the URL).

https://forum.nette.org/cs/34670-user-warning-form-was-submitted-but-there-are-no-associated-handlers