Send heartbeat ping once SSL/TLS tests finish
The URL is defined in GitHub Actions Repo Secrets in repo settings, and currently points to UptimeRobot.

`if: always()` makes sure the ping is sent no matter the result of the previous steps, we just want to be sure the tests are executed even if they fail. See https://docs.github.com/en/actions/learn-github-actions/workflow-syntax-for-github-actions#example-not-requiring-dependent-jobs-to-be-successful