Resolve PHPStan error
PHPStan says "Variable $slide might not be defined." although it will always be at this point. It's defined in the foreach loop but PDOException can also be thrown only later in the loop.

PhpStorm complains "Condition is always 'true' because '$slideNumber' is evaluated at this point" but let's leave it this way to make PHPStan satisfied.