Add a custom error presenter for Pulse
The goal is to have a lite error page which doesn't log to Nette's access.log because I get a lot of requests for some adult sites like
- `/passwords/storages/site/xxxvideo4[.]com/result/saranno%20famos`
- `/passwords/storages/site/beeg[.]faith`
My guess is that someone found that whatever they put after `site/` is echoed to the 404 page and they or their bot thought "oh cool this will help us for SEO" (because they didn't check properly?)

But I don't care and don't want those in my logs. It will still be in nginx error log but that's fine.

The HTML inspired by plain Apache "not found" page
https://github.com/apache/httpd/tree/trunk/docs/error