Reduce generated DI container size by exporting only needed metadata
The DI container on my local was reduced 117 kB → 110 kB which is still worth it.
Could limit it more by specifying classes used with `$container->getByType()` but there's quite a few of them and sometimes they're in extensions like the SRI one so I consider it too risky if not automated somehow.

https://doc.nette.org/en/dependency-injection/configuration#toc-di