Set wwwDir param for tests, too
If not set, the default is used and the default is calculated as a directory in which the executed test file is stored. So the wwwDir is different for each test file executed which results in many containers (one per dir) being created when running the full test suite which slows it down.

Follow-up to #27