Update spaze/sri-macros and spaze/phpstan-disallowed-calls
This is an extra commit because things got a little complicated here.

Due to changes spaze/phpstan-disallowed-calls 2.11.3 and particularly spaze/phpstan-disallowed-calls#159, more params can be detected and PHPStan started throwing errors in spaze/sri-macros/src/Config.php:126 saying "Calling hash_file() is forbidden". The type that got passed to hash_file() was `StringType`, not `ConstantStringType` and even though the values were the ones allowed, the param was rejected.

I had to change sri-macros to pass a union of `ConstantStringType`s and phpstan-disallowed-calls to support union types.

 - spaze/phpstan-disallowed-calls updated from v2.11.3 to v2.11.4 patch
   See changes: https://github.com/spaze/phpstan-disallowed-calls/compare/v2.11.3...v2.11.4
   Release notes: https://github.com/spaze/phpstan-disallowed-calls/releases/tag/v2.11.4

 - spaze/sri-macros updated from v2.0.0 to v2.0.1 patch
   See changes: https://github.com/spaze/sri-macros/compare/v2.0.0...v2.0.1
   Release notes: https://github.com/spaze/sri-macros/releases/tag/v2.0.1