Add well-known URL for changing passwords in admin
The redirect is done in Nette so that I can redirect to presenter:action and not a URL. If the redirect would be done in server config, then it's one more place to change if the URL changes.
Also not added as a one-way route to Nette because the redirection should be a temporary redirect while the one-way route redirect is a permanent one.

See the spec here https://www.w3.org/TR/change-password-url/