Test all targets in the Makefile are phony
A phony target is one that runs its recipe even if a file with the target's name already exists. I want all targets to be phony in case a file of the target's name suddenly appears which would prevent the recipe (and the test the recipe runs) to be executed.

Another option would be to check existing files whether they're not of the same name as the target names. Such check would need to run early enough before all tests.