Change what decides whether the email additional text container should be open
It was a `class=hidden` on the container itself, checking form component's value to see if it should be added or not. Now it's `class=expanded` on an element defined by `data-class-target`, a textarea in this case, where `class=expanded` is added in the form factory if needed.

CSS `:has()` is currently not supported by Firefox (it's behind a flag) but guess that's ok by me.

Fixes
```
 ------ ---------- ------------------------------------------------------------------------------------------------------------------------------------
  Line   Compiled   Admin/Presenters/templates/Emails/default.latte rendered from MichalSpacekCz\Admin\Presenters\EmailsPresenter::default
         line       See compiled template: /tmp/phpstan-latte/app/Admin/Presenters/templates/Emails/default.latte.ffc5a80775bd006bd5dcca69684ee311.php
 ------ ---------- ------------------------------------------------------------------------------------------------------------------------------------
  63     364        Cannot access offset 'additional' on Nette\Forms\Controls\BaseControl.
 ------ ---------- ------------------------------------------------------------------------------------------------------------------------------------
```