Send UPC-keys-as-text with a method, not a template
This is because an extra template means it could be displayed even without the action/render method in the presenter. Well, not in this case, because the router wouldn't allow it, but in general. See the note below the diagram here https://doc.nette.org/en/application/presenters#toc-life-cycle-of-presenter

I could move the template to some other directory but I think just building the string is faster and simpler.

Ref efabrica-team/phpstan-latte#415