Add default null values to "current /something/" properties
These will always be initialized in the `action*()` methods and so will always be available in the `createComponent*()` methods. But Psalm can't possibly know that. Still not sure I like it this way, it's there mostly for Psalm only but maybe it's ok.

Resolves errors like
```
ERROR: PropertyNotSetInConstructor - app/Admin/Presenters/TrainingsPresenter.php:44:30 - Property MichalSpacekCz\Admin\Presenters\TrainingsPresenter::$application is not defined in constructor of MichalSpacekCz\Admin\Presenters\TrainingsPresenter or in any private or final methods called in the constructor (see https://psalm.dev/074)
        private TrainingApplication $application;
```