Forbid `Nette\Forms\Form`, `createComponent...()` methods return the new `UiForm` anyway
It's still used in `TrainingControlsFactory` for rules e.g. `addRule(Form::EMAIL, ...)` because the methods in `TrainingControlsFactory::addAttendee()` don't always get just a form, but sometimes that's a container being passed into them. And while I could call `getForm()` on it, I think that would be an overkill just to get a constant from an otherwise disallowed class. That's why `TrainingControlsFactory.php` is in `allowIn`.