Duplicate mapping config from the main app to the PHPStan Latte extension, again
Also adds a check that the config is the same in all files. I'd need to add `unformatPresenterClass()` somewhere somehow if it wasn't the duplicated config. And while that's a copy/pasta, with the test it's acceptable.

Partially rolls back #84, see https://github.com/efabrica-team/phpstan-latte/pull/456 for a bit more details on what's going on.