Revert "loading" button caption when Nette's form validation error appears
the validation is triggered on form submit but at that moment, the button caption is already changed to "loading" by the click event handler. But because the preview frame never loads then, the "loading" caption would stay.

So we'll observe any child list changes of the BODY element and see if there's a form validation DIALOG element, and if yes, revert the caption.