Add CSP snippet for httpbin fetch from console for example
When I say that people can try for example `fetch('https://httpbin.org')` in their devtools, I expect people to open devtools on the page they're reading. So makes sense to allow it so it doesn't fail for them on CSP. I'll publish an article soon that may use this.