Adding assert()s for PHPStan to recognize $row types
Not fully sure I like it, there's still a very loose connection between the types in the assert and the database types, so if the type in the database changes, I'd need to update the assert as well. But previously, there was no connection between those two at all, so maybe it is better this way. Well, at least PHPStan doesn't now complain.

Got ~250 errors after upgrade to PHPStan 1.12.0, errors like

> Parameter #1 $... of method ... expects string, mixed given.