Remove version number from composer.json
`composer validate` now throws a general warning:
"The version field is present, it is recommended to leave it out if the package is published on Packagist."

Although the package is not published on Packagist, but here we are. Composer was throwing a different error before, and that's why the version number was added.

Reverts commit a70519114db536a3f91a2efb4d3bfbdd1a4ad5f6 #296