Show FourOhFourButFound output, if any, not just on 404s but also on existing content
This means for example that previously, `/?file=/etc/passwd` didn't trigger the FourOhFourButFound output because `/` is a valid existing URL, but now it will.

It uses $_SERVER because that contains the original requested URL, Nette's Request service has it already sanitized a bit and will not retain new lines for example, which some CVE exploits use. This means I can troll more folks, yay!