Remove the custom generic Gitleaks rule as it's not needed anymore
It was added in #76 to detect "dev" and "prod" keys, but I've since started to use prefixed encryption keys in #269 so the custom rule is not needed anymore. This is why you use prefixed keys, it makes things simple, and I like making things simple.

Also, the custom generic rule started throwing many false positives after I've upgraded to Gitleaks 8.23.1 for some reason.