Disallow all array_filter() calls without the callback param
Even though some calls are not flagged by phpstan-strict-rules when the 2nd param is missing (when the array consists of just arrays or something like that), let's just disallow all of the calls, mostly because I want a custom disallow message to remember with what I should replace the call. It was just one place to change the array_filter() call to Arrays::filterEmpty().