Ignore the "stream_get_contents returns false in specific conditions" error
And do it for PHPStan, as ignoring it in Psalm is a bit more complicated, and PHPStan will get it "fixed" soon, and then this ignore can be removed. I'd rather make Psalm not report the PossiblyFalseArgument when the `=== false` check is removed, but here we are.

And regarding the specific conditions: stream_get_contents() does not return FALSE unless an invalid position is provided. I don't provide any position so it cannot return false.