Set a different cache dir for CLI scripts and tests
To avoid having to set a username using `runuser` when cleaning stale sessions because the cache dir is owned by `www-user` while the clean script runs under `ubuntu`, and to avoid config conflicts when running tests because tests use different domains for example.

Modifying the `defaultExtensions` array is not something I'm super proud of, but seems to be the only way to change the cache dir (not the parent temp dir). It's also tested so should be mostly fine.

Semi-followup to #567