Add a Cron config file to clean stale sessions every 30 minutes
The time interval follows Ondřej Surý's PHP package interval.

The reason for not using a systemd timer is simplicity. I don't care much about logging as the script logs the result to database anyway.

I don't also mind if the script would miss the execution interval when the server would be down for example at that moment as stale sessions would be deleted when the script would be executed next, the 30-minute interval is quite short. Sessions also have their own TTL no matter how long the file (or the database row in my case) would physically exist, so there's no need to delete files/rows as soon as possible.

I could use `~/.crontab` (which then needs to be installed with `cron .crontab`) but I need the script to run under `www-user` and that's a bit complicated to to in another user's crontab (must do something like `sudo runuser -u www-user ...`). Another reason is that I wouldn't be able to have a file in the repo that can only be included or symlinked somewhere. That can only be done when I'm using `/etc/cron.d`, `/etc/cron.hourly` and similar. I've chosen `/etc/cron.d` because you can specify a username which the script will use to run under.

If using systemd, you can put user timers in `~/.config/systemd/user/`. But those are only executed when the user is logged in, which is not the case for servers. I could enable the linger mode which would then execute timers even when the user is not logged in, but I don't want to change a system configuration. And I don't need to use systemd as I've already explained.