Run clean_sessions.php from user ubuntu crontab
Running scripts from /etc/cron.d requires the script to be owned by root, which clean_sessions.php is obviously not. So it would require a tiny wrapper that would need to be set up manually anyway, and in that case, I can already go with setting the crontab manually, which is what I've went with. Another advantage is that everything is set up in ubuntu's home dir and not in /etc.

Follow-up to #567, partially supersedes #572.