Serve `security.txt` files by the app instead of nginx
That means that the request will always redirect `/.well-known/security.txt` to the "main site" (with `www`, or without).

Because spaze/security-txt has introduced validation of the `Canonical` field in https://github.com/spaze/security-txt/pull/44 and when it doesn't match, a warning is generated.

Another option would be to add extra `Canonical` (e.g. with `www`) to the `security.txt` file, or to redirect directly in nginx, but that would mean having redirect rules, like is this domain used with or without `www`, in one more place. Let's see if this is a good idea.