Use the PHP 8.5's Uri extension for parsing URLs instead of `parse_url()`
Use the WHATWG parser as that's what browsers follow, and what I'm doing here is mostly about browsers and websites and things like that.

It would also be cool if frameworks started using the extension so that all URL parsing is done the same way, but I guess we'll need to wait.

https://www.php.net/releases/8.5/en.php#new-uri-extension
https://thephp.foundation/blog/2025/10/10/php-85-uri-extension/